# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 巡梦 Dream Journal API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CharacterLanguage = typeof CharacterLanguage[keyof typeof CharacterLanguage];


export const CharacterLanguage = {
  zh: 'zh',
  en: 'en',
} as const;

export interface Character {
  id: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  role: string;
  personality: string[];
  speakingStyle: string;
  relationship: string;
  language: CharacterLanguage;
  voiceType: string;
  systemPrompt: string;
  isActive: boolean;
  createdAt: string;
}

export type CharacterInputLanguage = typeof CharacterInputLanguage[keyof typeof CharacterInputLanguage];


export const CharacterInputLanguage = {
  zh: 'zh',
  en: 'en',
} as const;

export interface CharacterInput {
  name: string;
  /** @nullable */
  avatar?: string | null;
  role: string;
  personality: string[];
  speakingStyle: string;
  relationship: string;
  language: CharacterInputLanguage;
  voiceType: string;
  systemPrompt: string;
  isActive?: boolean;
}

export type CharacterUpdateLanguage = typeof CharacterUpdateLanguage[keyof typeof CharacterUpdateLanguage];


export const CharacterUpdateLanguage = {
  zh: 'zh',
  en: 'en',
} as const;

export interface CharacterUpdate {
  name?: string;
  /** @nullable */
  avatar?: string | null;
  role?: string;
  personality?: string[];
  speakingStyle?: string;
  relationship?: string;
  language?: CharacterUpdateLanguage;
  voiceType?: string;
  systemPrompt?: string;
  isActive?: boolean;
}

export type DreamMood = typeof DreamMood[keyof typeof DreamMood];


export const DreamMood = {
  calm: 'calm',
  anxious: 'anxious',
  happy: 'happy',
  fearful: 'fearful',
  confused: 'confused',
  nostalgic: 'nostalgic',
  strange: 'strange',
} as const;

export type DreamClarity = typeof DreamClarity[keyof typeof DreamClarity];


export const DreamClarity = {
  vague: 'vague',
  moderate: 'moderate',
  vivid: 'vivid',
} as const;

export interface Dream {
  id: string;
  title: string;
  content: string;
  mood: DreamMood;
  clarity: DreamClarity;
  isRecurring: boolean;
  createdAt: string;
  /** @nullable */
  characterId?: string | null;
  /** @nullable */
  summary?: string | null;
  symbols?: string[];
  /** @nullable */
  emotionAnalysis?: string | null;
  /** @nullable */
  possibleConnection?: string | null;
  /** @nullable */
  companionReply?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type DreamInputMood = typeof DreamInputMood[keyof typeof DreamInputMood];


export const DreamInputMood = {
  calm: 'calm',
  anxious: 'anxious',
  happy: 'happy',
  fearful: 'fearful',
  confused: 'confused',
  nostalgic: 'nostalgic',
  strange: 'strange',
} as const;

export type DreamInputClarity = typeof DreamInputClarity[keyof typeof DreamInputClarity];


export const DreamInputClarity = {
  vague: 'vague',
  moderate: 'moderate',
  vivid: 'vivid',
} as const;

export interface DreamInput {
  title: string;
  content: string;
  mood: DreamInputMood;
  clarity: DreamInputClarity;
  isRecurring: boolean;
  /** @nullable */
  characterId?: string | null;
  /** @nullable */
  summary?: string | null;
  symbols?: string[];
  /** @nullable */
  emotionAnalysis?: string | null;
  /** @nullable */
  possibleConnection?: string | null;
  /** @nullable */
  companionReply?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type DreamUpdateMood = typeof DreamUpdateMood[keyof typeof DreamUpdateMood];


export const DreamUpdateMood = {
  calm: 'calm',
  anxious: 'anxious',
  happy: 'happy',
  fearful: 'fearful',
  confused: 'confused',
  nostalgic: 'nostalgic',
  strange: 'strange',
} as const;

export type DreamUpdateClarity = typeof DreamUpdateClarity[keyof typeof DreamUpdateClarity];


export const DreamUpdateClarity = {
  vague: 'vague',
  moderate: 'moderate',
  vivid: 'vivid',
} as const;

export interface DreamUpdate {
  title?: string;
  content?: string;
  mood?: DreamUpdateMood;
  clarity?: DreamUpdateClarity;
  isRecurring?: boolean;
  /** @nullable */
  characterId?: string | null;
  /** @nullable */
  summary?: string | null;
  symbols?: string[];
  /** @nullable */
  emotionAnalysis?: string | null;
  /** @nullable */
  possibleConnection?: string | null;
  /** @nullable */
  companionReply?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type DreamStatsMoodBreakdown = {[key: string]: number};

export type DreamStatsClarityBreakdown = {[key: string]: number};

export interface DreamStats {
  total: number;
  moodBreakdown: DreamStatsMoodBreakdown;
  clarityBreakdown: DreamStatsClarityBreakdown;
  recurringCount: number;
  recentSymbols: string[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  characterId?: string | null;
  /** @nullable */
  dreamId?: string | null;
  createdAt: string;
}

export interface AiOrganizeInput {
  content: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  characterSystemPrompt?: string | null;
}

export interface AiOrganizeResult {
  summary: string;
  symbols: string[];
  emotionAnalysis: string;
  possibleConnection: string;
  companionReply: string;
  isMock: boolean;
}

export type AiChatHistoryItemRole = typeof AiChatHistoryItemRole[keyof typeof AiChatHistoryItemRole];


export const AiChatHistoryItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AiChatHistoryItem {
  role: AiChatHistoryItemRole;
  content: string;
}

export interface AiChatInput {
  message: string;
  /** @nullable */
  characterSystemPrompt?: string | null;
  /** @nullable */
  dreamContext?: string | null;
  history: AiChatHistoryItem[];
}

export interface AiChatResult {
  reply: string;
  isMock: boolean;
}

export interface AiImageInput {
  imageBase64: string;
  mimeType?: string;
}

export interface AiImageResult {
  description: string;
  recognizedText: string;
  dreamElements: string[];
  suggestedTitle: string;
  draftContent: string;
  isMock: boolean;
}

export type AiSettingsMode = typeof AiSettingsMode[keyof typeof AiSettingsMode];


export const AiSettingsMode = {
  mock: 'mock',
  real: 'real',
} as const;

export interface AiSettings {
  mode: AiSettingsMode;
  hasApiKey: boolean;
  modelName: string;
}

export type GetChatHistoryParams = {
characterId?: string;
dreamId?: string;
limit?: number;
};

export type ClearChatHistoryParams = {
characterId?: string;
};