# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: 巡梦 Dream Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List all character profiles
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new character profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/active:
    get:
      operationId: getActiveCharacter
      tags: [characters]
      summary: Get the currently active character
      responses:
        "200":
          description: Active character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: No active character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /characters/{id}/activate:
    post:
      operationId: activateCharacter
      tags: [characters]
      summary: Set a character as the active companion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat message history
      parameters:
        - name: characterId
          in: query
          required: false
          schema:
            type: string
        - name: dreamId
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /chat/clear:
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history for a character
      parameters:
        - name: characterId
          in: query
          required: false
          schema:
            type: string
      responses:
        "204":
          description: Cleared
  /ai/organize:
    post:
      operationId: aiOrganizeDream
      tags: [ai]
      summary: Use AI to organize and analyze a dream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiOrganizeInput"
      responses:
        "200":
          description: AI organized dream analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiOrganizeResult"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI companion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResult"
  /ai/recognize-image:
    post:
      operationId: aiRecognizeImage
      tags: [ai]
      summary: Use AI to recognize and extract dream content from an image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiImageInput"
      responses:
        "200":
          description: AI image recognition result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiImageResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Character:
      type: object
      required:
        - id
        - name
        - role
        - personality
        - speakingStyle
        - relationship
        - language
        - voiceType
        - systemPrompt
        - isActive
        - createdAt
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
        personality:
          type: array
          items:
            type: string
        speakingStyle:
          type: string
        relationship:
          type: string
        language:
          type: string
          enum: [zh, en]
        voiceType:
          type: string
        systemPrompt:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    CharacterInput:
      type: object
      required:
        - name
        - role
        - personality
        - speakingStyle
        - relationship
        - language
        - voiceType
        - systemPrompt
      properties:
        name:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
        personality:
          type: array
          items:
            type: string
        speakingStyle:
          type: string
        relationship:
          type: string
        language:
          type: string
          enum: [zh, en]
        voiceType:
          type: string
        systemPrompt:
          type: string
        isActive:
          type: boolean
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
        personality:
          type: array
          items:
            type: string
        speakingStyle:
          type: string
        relationship:
          type: string
        language:
          type: string
          enum: [zh, en]
        voiceType:
          type: string
        systemPrompt:
          type: string
        isActive:
          type: boolean
...
    ChatMessage:
      type: object
      required:
        - id
        - role
        - content
        - createdAt
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        characterId:
          type: ["string", "null"]
        dreamId:
          type: ["string", "null"]
        createdAt:
          type: string
    AiOrganizeInput:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        title:
          type: ["string", "null"]
        characterSystemPrompt:
          type: ["string", "null"]
    AiOrganizeResult:
      type: object
      required:
        - summary
        - symbols
        - emotionAnalysis
        - possibleConnection
        - companionReply
        - isMock
      properties:
        summary:
          type: string
        symbols:
          type: array
          items:
            type: string
        emotionAnalysis:
          type: string
        possibleConnection:
          type: string
        companionReply:
          type: string
        isMock:
          type: boolean
    AiChatInput:
      type: object
      required:
        - message
        - history
      properties:
        message:
          type: string
        characterSystemPrompt:
          type: ["string", "null"]
        dreamContext:
          type: ["string", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/AiChatHistoryItem"
    AiChatHistoryItem:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    AiChatResult:
      type: object
      required:
        - reply
        - isMock
      properties:
        reply:
          type: string
        isMock:
          type: boolean
    AiImageInput:
      type: object
      required:
        - imageBase64
      properties:
        imageBase64:
          type: string
        mimeType:
          type: string
    AiImageResult:
      type: object
      required:
        - description
        - recognizedText
        - dreamElements
        - suggestedTitle
        - draftContent
        - isMock
      properties:
        description:
          type: string
        recognizedText:
          type: string
        dreamElements:
          type: array
          items:
            type: string
        suggestedTitle:
          type: string
        draftContent:
          type: string
        isMock:
          type: boolean
...