# lib/api-zod/src/generated/types/aiChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 巡梦 Dream Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { AiChatHistoryItem } from './aiChatHistoryItem';

export interface AiChatInput {
  message: string;
  /** @nullable */
  characterSystemPrompt?: string | null;
  /** @nullable */
  dreamContext?: string | null;
  history: AiChatHistoryItem[];
}