# lib/api-zod/src/generated/types/aiImageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 巡梦 Dream Journal API
 * OpenAPI spec version: 0.2.0
 */

export interface AiImageInput {
  imageBase64: string;
  mimeType?: string;
}