# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 巡梦 Dream Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { CharacterLanguage } from './characterLanguage';

export interface Character {
  id: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  role: string;
  personality: string[];
  speakingStyle: string;
  relationship: string;
  language: CharacterLanguage;
  voiceType: string;
  systemPrompt: string;
  isActive: boolean;
  createdAt: string;
}